Align export analysis source layout with webpack
