+++ crates/unpack_core/src/flag_dependency_usage_plugin.rs
@@ -1,26 +1,8 @@
 use std::collections::{BTreeSet, HashMap};
 
 use crate::{
-    Compilation, Dependency, ModuleHandle,
-    hooks::{CompilationHooks, CompilerHooks},
+    Compilation, Dependency, ModuleHandle, compilation::CompilationHooks, compiler::CompilerHooks,
 };
-
-pub(crate) struct FlagDependencyExportsPlugin;
-
-impl FlagDependencyExportsPlugin {
-    pub fn apply(&self, hooks: &mut CompilerHooks) {
-        hooks.compilation.tap(
-            "FlagDependencyExportsPlugin",
-            |compilation_hooks: &mut CompilationHooks| {
-                compilation_hooks
-                    .finish_modules
-                    .tap("FlagDependencyExportsPlugin", |compilation| {
-                        Box::pin(async move { flag_dependency_exports(compilation) })
-                    });
-            },
-        );
-    }
-}
 
 pub(crate) struct FlagDependencyUsagePlugin;
 
@@ -34,20 +16,6 @@ impl FlagDependencyUsagePlugin {
                     .tap("FlagDependencyUsagePlugin", flag_dependency_usage);
             },
         );
-    }
-}
-
-fn flag_dependency_exports(compilation: &mut Compilation) {
-    let handles = compilation
-        .module_graph()
-        .modules()
-        .iter()
-        .map(|module| module.handle())
-        .collect::<Vec<_>>();
-    for handle in handles {
-        if let Some(module) = compilation.module_graph_mut().module_mut(handle) {
-            module.analyze_provided_exports();
-        }
     }
 }
