+++ crates/unpack_core/src/lib.rs
@@ -10,9 +10,9 @@ mod compilation;
 mod compiler;
 mod dependency;
 mod error;
-mod export_analysis;
 mod exports_info;
-mod hooks;
+mod flag_dependency_exports_plugin;
+mod flag_dependency_usage_plugin;
 mod id_assignment;
 mod loader;
 mod logging;
