+++ docs/adr/0113-align-internal-model-with-webpack-responsibilities.md
@@ -2,4 +2,8 @@
 
 Unpack will align internal data structures and algorithms with webpack's names, phase ordering, and responsibility boundaries where practical: concepts such as `Compilation`, `ModuleGraph`, `ChunkGraph`, `ChunkGroup`, `Entrypoint`, `NormalModuleFactory`, `Dependency`, `DependenciesBlock`, `AsyncDependenciesBlock`, `DependencyTemplate`, and `RuntimeRequirement` should remain recognizable. The Rust implementation may use Rust-native traits, enums, ownership, and concurrency instead of copying webpack's JavaScript class hierarchy, mutable object layout, hook storage, or `AsyncQueue` API unless those shapes become necessary to match exposed plugin or loader behavior.
 
-Source layout will follow the same rule for implemented responsibilities. Distinct webpack concepts and algorithms such as `Chunk`, `ChunkGroup`, `ChunkGraph`, and `buildChunkGraph` should remain separately locatable, and subsystem directories such as `dependencies`, `runtime`, and `cache` should be introduced as those implementations grow. This does not require one Rust file per JavaScript class: closed class families may remain enums, dense graph handles may index `Vec` storage, closed requirement sets may use bit sets, and hot-path set algebra may use masks instead of JavaScript-shaped `Set` or `Map` objects. Dense Rust keys use `*Handle` or `*Index` names rather than taking webpack's Module ID and Chunk ID names away from generated output identity.
+Source layout will follow the same rule for implemented responsibilities. Unless a concrete Rust language, ownership, performance, or build-system constraint requires otherwise, Unpack's directory hierarchy and file organization must map directly to webpack's `lib/` layout. A webpack class or plugin implemented by Unpack should have its own correspondingly named Rust module at the equivalent source level. Implementation-specific aggregate files must not combine webpack units that are separate files upstream, and hooks owned by webpack's `Compiler` or `Compilation` belong under the corresponding Rust module boundary.
+
+Any layout deviation must be deliberate and documented with the special constraint that requires it. Rust naming conventions such as snake_case filenames are not deviations. Moving a webpack unit to a different directory, combining multiple upstream files into one module, or splitting one upstream responsibility across unrelated directories is a deviation and requires an ADR or an update to the relevant implementation-differences document.
+
+Distinct webpack concepts and algorithms such as `Chunk`, `ChunkGroup`, `ChunkGraph`, and `buildChunkGraph` should remain separately locatable, and subsystem directories such as `dependencies`, `runtime`, and `cache` should be introduced as those implementations grow. Closed class families may remain enums, dense graph handles may index `Vec` storage, closed requirement sets may use bit sets, and hot-path set algebra may use masks instead of JavaScript-shaped `Set` or `Map` objects only when that Rust-specific representation is the documented reason for the different file boundary. Dense Rust keys use `*Handle` or `*Index` names rather than taking webpack's Module ID and Chunk ID names away from generated output identity.
